implement load data from csv and

+++ src/client_side/ui/models/Model.java
@@ -1,6 +1,5 @@
 package client_side.ui.models;
 
-import javafx.beans.Observable;
 import utility.EmptyObservable;
 
 import java.io.IOException;
@@ -15,7 +14,7 @@ public interface Model {
     void runScript(String script) throws IllegalAccessException;
     void stopScript();
 
-    void calculatePath(String ip, int port, double[][] heights, int sourceRow, int sourceCol, int destRow, int destCol);
+    void calculatePath(String ip, int port, int[][] heights, int sourceRow, int sourceCol, int destRow, int destCol);
     String getPath();
     EmptyObservable getPathDoneObservable();
